Revert "Liambly/atc ua generalization"
