+++ lib/user_agent/gem_version.rb
@@ -2,6 +2,6 @@
 
 class UserAgent
   module GemVersion
-    VERSION = '1.5.2'
+    VERSION = '1.5.1'
   end
 end
