+++ lib/user_agent/operating_systems/darwin.rb
@@ -55,7 +55,7 @@ class UserAgent
         '21.2.0' => '15.2',   # also 15.2.1
         '21.3.0' => '15.3',   # also 15.3.1
         '21.4.0' => '15.4',   # also 15.4.1
-        '21.5.0' => '15.5',
+        '21.5.0' => '15.5',   
         '21.6.0' => '15.6',   # also 15.6.1, 15.7 and 15.7.1
         '22.0.0' => '16.0.x',
         '22.1.0' => '16.1',
